+++ vector/src/main/java/im/vector/app/features/roommemberprofile/RoomMemberProfileFragment.kt
@@ -43,6 +43,7 @@ import im.vector.app.core.extensions.setTextOrHide
 import im.vector.app.core.platform.StateView
 import im.vector.app.core.platform.VectorBaseFragment
 import im.vector.app.core.utils.startSharePlainTextIntent
+import im.vector.app.databinding.DialogBaseEditTextBinding
 import im.vector.app.databinding.DialogShareQrCodeBinding
 import im.vector.app.databinding.FragmentMatrixProfileBinding
 import im.vector.app.databinding.ViewStubRoomMemberProfileHeaderBinding
@@ -50,9 +51,11 @@ import im.vector.app.features.crypto.verification.VerificationBottomSheet
 import im.vector.app.features.home.AvatarRenderer
 import im.vector.app.features.home.room.detail.RoomDetailPendingAction
 import im.vector.app.features.home.room.detail.RoomDetailPendingActionStore
+import im.vector.app.features.home.room.detail.timeline.helper.MatrixItemColorProvider
 import im.vector.app.features.roommemberprofile.devices.DeviceListBottomSheet
 import im.vector.app.features.roommemberprofile.powerlevel.EditPowerLevelDialogs
 import kotlinx.parcelize.Parcelize
+import org.matrix.android.sdk.api.session.accountdata.UserAccountDataTypes
 import org.matrix.android.sdk.api.session.room.powerlevels.Role
 import org.matrix.android.sdk.api.util.MatrixItem
 import javax.inject.Inject
@@ -67,7 +70,8 @@ class RoomMemberProfileFragment @Inject constructor(
         val viewModelFactory: RoomMemberProfileViewModel.Factory,
         private val roomMemberProfileController: RoomMemberProfileController,
         private val avatarRenderer: AvatarRenderer,
-        private val roomDetailPendingActionStore: RoomDetailPendingActionStore
+        private val roomDetailPendingActionStore: RoomDetailPendingActionStore,
+        private val matrixItemColorProvider: MatrixItemColorProvider
 ) : VectorBaseFragment<FragmentMatrixProfileBinding>(),
         RoomMemberProfileController.Callback {
 
@@ -199,6 +203,7 @@ class RoomMemberProfileFragment @Inject constructor(
                 headerViews.memberProfileIdView.text = userMatrixItem.id
                 val bestName = userMatrixItem.getBestName()
                 headerViews.memberProfileNameView.text = bestName
+                headerViews.memberProfileNameView.setTextColor(matrixItemColorProvider.getColor(userMatrixItem))
                 views.matrixProfileToolbarTitleView.text = bestName
                 avatarRenderer.render(userMatrixItem, headerViews.memberProfileAvatarView)
                 avatarRenderer.render(userMatrixItem, views.matrixProfileToolbarAvatarImageView)
@@ -236,6 +241,9 @@ class RoomMemberProfileFragment @Inject constructor(
 
                 headerViews.memberProfileAvatarView.setOnClickListener { view ->
                     onAvatarClicked(view, userMatrixItem)
+                }
+                headerViews.memberProfileNameView.setOnClickListener { _ ->
+                    onProfileNameClicked(userMatrixItem)
                 }
                 views.matrixProfileToolbarAvatarImageView.setOnClickListener { view ->
                     onAvatarClicked(view, userMatrixItem)
@@ -321,6 +329,40 @@ class RoomMemberProfileFragment @Inject constructor(
 
     private fun onAvatarClicked(view: View, userMatrixItem: MatrixItem) {
         navigator.openBigImageViewer(requireActivity(), view, userMatrixItem)
+    }
+
+    private fun onProfileNameClicked(userMatrixItem: MatrixItem) {
+        val inflater = requireActivity().layoutInflater
+        val layout = inflater.inflate(R.layout.dialog_base_edit_text, null)
+        val views = DialogBaseEditTextBinding.bind(layout)
+        val session = injector().activeSessionHolder().getActiveSession()
+        val overrideColorsSetting = session.getAccountDataEvent(UserAccountDataTypes.TYPE_OVERRIDE_COLORS)
+        val overrideColorSpecs = overrideColorsSetting?.content?.toMap().orEmpty()
+        val overrideColorSpec = overrideColorSpecs[userMatrixItem.id]?.toString()
+        views.editText.setText(overrideColorSpec)
+        views.editText.hint = "#000000"
+
+        AlertDialog.Builder(requireActivity())
+                .setTitle(R.string.room_member_override_color)
+                .setView(layout)
+                .setPositiveButton(R.string.ok) { _, _ ->
+                    val newOverrideColorSpec = views.editText.text.toString()
+                    if (newOverrideColorSpec != overrideColorSpec) {
+                        val newOverrideColorSpecs = overrideColorSpecs.toMutableMap()
+                        if (matrixItemColorProvider.setOverrideColor(userMatrixItem.id, newOverrideColorSpec)) {
+                            newOverrideColorSpecs[userMatrixItem.id] = newOverrideColorSpec
+                        } else {
+                            newOverrideColorSpecs.remove(userMatrixItem.id)
+                        }
+                        session.updateAccountData(
+                                type = UserAccountDataTypes.TYPE_OVERRIDE_COLORS,
+                                content = newOverrideColorSpecs
+                        )
+                        headerViews.memberProfileNameView.setTextColor(matrixItemColorProvider.getColor(userMatrixItem))
+                    }
+                }
+                .setNegativeButton(R.string.cancel, null)
+                .show()
     }
 
     override fun onEditPowerLevel(currentRole: Role) {